+++ peerprep-fe/src/app/collaboration/components/CollaborationEditor.tsx
@@ -32,6 +32,7 @@ import {
   AlertDialogTrigger,
 } from '@/components/ui/alert-dialog';
 import { useRouter } from 'next/navigation';
+import { stringToColor } from '@/lib/utils';
 
 interface CollaborationEditorProps {
   language: string;
@@ -59,15 +60,6 @@ const CollaborationEditor = ({
   const { toast } = useToast();
   const { clearLastMatchId } = useCollaborationStore();
   const router = useRouter();
-
-  const stringToColor = (str: string) => {
-    let hash = 0;
-    for (let i = 0; i < str.length; i++) {
-      hash = str.charCodeAt(i) + ((hash << 5) - hash);
-    }
-    const color = Math.abs(hash).toString(16).substring(0, 6);
-    return `#${'0'.repeat(6 - color.length)}${color}`;
-  };
 
   const handleEditorMount: OnMount = (editor) => {
     if (!matchId) {
