+++ peerprep-fe/src/lib/utils.ts
@@ -65,3 +65,12 @@ export const validateUsername = (username: string): string[] => {
   }
   return errors;
 };
+
+export const stringToColor = (str: string) => {
+  let hash = 0;
+  for (let i = 0; i < str.length; i++) {
+    hash = str.charCodeAt(i) + ((hash << 5) - hash);
+  }
+  const color = Math.abs(hash).toString(16).substring(0, 6);
+  return `#${'0'.repeat(6 - color.length)}${color}`;
+};
